requirements.txt
astroid==2.3.3
colorama==0.4.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==10.0.0
pylint==2.4.4
six==1.13.0
typed-ast==1.5.5
wrapt==1.11.2
